rehomed various tests
