+++ tests/json_test.h
@@ -9,6 +9,8 @@ namespace tests {
 void JsonDefaultTest(const std::string& tests_data_path);
 void JsonEnumsTest(const std::string& tests_data_path);
 void JsonOptionalTest(const std::string& tests_data_path, bool default_scalars);
+void ParseIncorrectMonsterJsonTest(const std::string& tests_data_path);
+void JsonUnsortedArrayTest();
 
 }  // namespace tests
 }  // namespace flatbuffers
